try to test low res data with a second test datsets
